requirements.txt
asgiref==3.8.1
boto3==1.40.45
botocore==1.40.45
distlib==0.3.6
dj-database-url==3.0.1
Django==5.1.2
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
filelock==3.9.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
platformdirs==3.1.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
s3transfer==0.14.0
Serializer==0.2.1
simplejson==3.20.2
six==1.17.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.5.0
virtualenv==20.21.0
whitenoise==6.11.0
